feat: add partners and projects page

+++ backend/controllers/partners.js
@@ -9,18 +9,18 @@ const filePath = path.join(__dirname, "../assets/images/partners");
 
 const getPartners = async (request, response) => {
 	try {
-		const { _id: partnerID, page, limit } = request.query;
+		const { _id: itemID, page, limit } = request.query;
 
-		if (!partnerID && (!page || !limit)) {
+		if (!itemID && (!page || !limit)) {
 			return sendJsonResponse(response, HTTP_STATUS_CODES.BAD_REQUEST, false, "Missing parameters!", null);
 		}
 
-		const dbPartners = await Partners.find(partnerID ? { _id: partnerID } : {})
+		const dbPayload = await Partners.find(itemID ? { _id: itemID } : {})
 			.limit(limit)
 			.skip(page && (page - 1) * limit);
 
-		if (dbPartners.length > 0) {
-			return sendJsonResponse(response, HTTP_STATUS_CODES.OK, true, "Record Found!", partnerID ? dbPartners[0] : dbPartners);
+		if (dbPayload.length > 0) {
+			return sendJsonResponse(response, HTTP_STATUS_CODES.OK, true, "Record Found!", itemID ? dbPayload[0] : dbPayload);
 		} else {
 			return sendJsonResponse(response, HTTP_STATUS_CODES.NOTFOUND, false, "Record not Found!", null);
 		}
@@ -31,9 +31,9 @@ const getPartners = async (request, response) => {
 
 const getPartnerImage = async (request, response) => {
 	try {
-		const { filename, width } = request.query;
+		const { filename, width, mimetype } = request.query;
 
-		if (!filename) {
+		if (!filename || !mimetype) {
 			return sendJsonResponse(response, HTTP_STATUS_CODES.BAD_REQUEST, false, "Missing parameters!", null);
 		}
 
@@ -41,10 +41,11 @@ const getPartnerImage = async (request, response) => {
 		const isFileExists = fs.existsSync(fileFullPath);
 
 		const sourceFile = fs.readFileSync(isFileExists ? fileFullPath : placeholderImage);
-		const optimizedImage = width ? await sharp(sourceFile).resize(parseInt(width)).toBuffer() : sourceFile;
+		const optimizedImage =
+			mimetype.startsWith("image") && width ? await sharp(sourceFile).resize(parseInt(width)).toBuffer() : sourceFile;
 
 		response.writeHead(200, {
-			"Content-Type": "image/webp",
+			"Content-Type": mimetype,
 		});
 
 		response.end(optimizedImage);
@@ -63,26 +64,29 @@ const createPartner = async (request, response) => {
 			return sendJsonResponse(response, HTTP_STATUS_CODES.BAD_REQUEST, false, "Missing parameters!", null);
 		}
 
-		for (let file of files) {
-			const webpImage = await convertImageToWebp(file);
-			const generatedFileName = generateUniqueFileName(webpImage, filePath);
+		if (files.length) {
+			let file = files[0];
 
+			if (file.mimetype.startsWith("image")) file = await convertImageToWebp(file);
+
+			const generatedFileName = generateUniqueFileName(file, filePath);
 			const fileFullPath = path.join(filePath, generatedFileName);
-			payload[file.fieldname] = generatedFileName;
 
-			await fs.promises.writeFile(fileFullPath, webpImage.buffer);
+			await fs.promises.writeFile(fileFullPath, file.buffer);
+
+			payload.media = { mimetype: file.mimetype, filename: generatedFileName };
 		}
 
-		const partner = new Partners({
+		const dbNewRecordObject = new Partners({
 			...payload,
 			createdBy: authenticatingUserID,
 			updatedBy: authenticatingUserID,
 		});
 
-		const newPartner = await partner.save();
+		const createdPayload = await dbNewRecordObject.save();
 
-		if (newPartner) {
-			return sendJsonResponse(response, HTTP_STATUS_CODES.OK, true, "Record created::success", newPartner);
+		if (createdPayload) {
+			return sendJsonResponse(response, HTTP_STATUS_CODES.OK, true, "Record created::success", createdPayload);
 		} else {
 			return sendJsonResponse(response, HTTP_STATUS_CODES.INTERNAL_SERVER_ERROR, false, "Record created::failure", null);
 		}
@@ -101,33 +105,36 @@ const updatePartner = async (request, response) => {
 			return sendJsonResponse(response, HTTP_STATUS_CODES.BAD_REQUEST, false, "Missing parameters!", null);
 		}
 
-		const dbPartner = await partners.findOne({ _id: payload._id });
+		const dbPayload = await Partners.findOne({ _id: payload._id });
 
 		if (files.length) {
-			for (let file of files) {
-				const webpImage = await convertImageToWebp(file);
-				const generatedFileName = generateUniqueFileName(webpImage, filePath);
+			let file = files[0];
 
-				const fileFullPath = path.join(filePath, generatedFileName);
+			if (file.mimetype.startsWith("image")) file = await convertImageToWebp(file);
 
-				const existingFilePath = path.join(filePath, dbPartner[webpImage.fieldname]);
+			const generatedFileName = generateUniqueFileName(file, filePath);
+			const fileFullPath = path.join(filePath, generatedFileName);
+
+			if (dbPayload?.media?.filename) {
+				const existingFilePath = path.join(filePath, dbPayload.media.filename);
+
 				const isThereExistingFile = fs.existsSync(existingFilePath);
 				if (isThereExistingFile) await fs.promises.unlink(existingFilePath);
-
-				await fs.promises.writeFile(fileFullPath, webpImage.buffer);
-
-				payload[file.fieldname] = generatedFileName;
 			}
+
+			await fs.promises.writeFile(fileFullPath, file.buffer);
+
+			payload.media = { mimetype: file.mimetype, filename: generatedFileName };
 		}
 
-		const updatedPartner = await Partners.findOneAndUpdate(
+		const updatedPayload = await Partners.findOneAndUpdate(
 			{ _id: payload._id },
 			{ $set: { ...payload, updatedBy: authenticatingUserID } },
 			{ new: true },
 		);
 
-		if (updatedPartner) {
-			return sendJsonResponse(response, HTTP_STATUS_CODES.OK, true, "Record updated::success", updatedPartner);
+		if (updatedPayload) {
+			return sendJsonResponse(response, HTTP_STATUS_CODES.OK, true, "Record updated::success", updatedPayload);
 		} else {
 			return sendJsonResponse(response, HTTP_STATUS_CODES.INTERNAL_SERVER_ERROR, false, "Record updated::failure", null);
 		}
@@ -138,21 +145,23 @@ const updatePartner = async (request, response) => {
 
 const deletePartner = async (request, response) => {
 	try {
-		const { _id: partnerID } = request.query;
+		const { _id: itemID } = request.query;
 
-		if (!partnerID) {
+		if (!itemID) {
 			return sendJsonResponse(response, HTTP_STATUS_CODES.BAD_REQUEST, false, "Missing parameters!", null);
 		}
 
-		const deletedPartner = await Partners.findOneAndDelete({ _id: partnerID }, { new: true });
+		const deletedPayload = await Partners.findOneAndDelete({ _id: itemID }, { new: true });
 
-		if (deletedPartner) {
-			const fileFullPath = path.join(filePath, deletedPartner.featuredImage);
+		if (deletedPayload) {
+			if (deletedPayload?.media?.filename) {
+				const fileFullPath = path.join(filePath, deletedPayload.media.filename);
 
-			const isfileExists = fs.existsSync(fileFullPath);
-			if (isfileExists) await fs.promises.unlink(fileFullPath);
+				const isfileExists = fs.existsSync(fileFullPath);
+				if (isfileExists) await fs.promises.unlink(fileFullPath);
+			}
 
-			return sendJsonResponse(response, HTTP_STATUS_CODES.OK, true, "Record delete::success", deletedPartner);
+			return sendJsonResponse(response, HTTP_STATUS_CODES.OK, true, "Record delete::success", deletedPayload);
 		} else {
 			return sendJsonResponse(response, HTTP_STATUS_CODES.INTERNAL_SERVER_ERROR, false, "Record delete::failure", null);
 		}
